CLEANUP: move to native package extensions in julia 1.9
