+++ Project.toml
@@ -26,14 +26,29 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StructEquality = "6ec83bb0-ed9f-11e9-3b4c-2b04cb4e219c"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
+[weakdeps]
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
+MetaGraphs = "626554b9-1ddb-594c-aa3c-2596fe9399a5"
+TikzPictures = "37f6aa50-8035-52d0-81c2-5a1d08754b2d"
+
+[extensions]
+CatlabDataFramesExt = "DataFrames"
+CatlabGraphsExt = "Graphs"
+CatlabMetaGraphsExt = "MetaGraphs"
+CatlabTikzPicturesExt = "TikzPictures"
+
 [compat]
 Colors = "0.12"
 CompTime = "0.1"
 Compose = "0.7, 0.8, 0.9"
+DataFrames = "1"
 DataStructures = "0.17, 0.18"
 GeneralizedGenerated = "0.2, 0.3"
+Graphs = "1"
 JSON = "0.20, 0.21"
 LightXML = "0.8, 0.9"
+MetaGraphs = "0.7"
 MLStyle = "0.4"
 PrettyTables = "^2"
 Reexport = "^1"
@@ -41,13 +56,10 @@ Requires = "^1"
 StaticArrays = "0.12, 1.0"
 StructEquality = "2.1"
 Tables = "^1"
+TikzPictures = "3"
 julia = "1.6"
 
 [extras]
 Convex = "f65535da-76fb-5f13-bab9-19810c17039a"
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 Graphviz_jll = "3c863552-8265-54e4-a6dc-903eb78fde85"
-MetaGraphs = "626554b9-1ddb-594c-aa3c-2596fe9399a5"
 SCS = "c946c3f1-0d1f-5ce8-9dea-7daa1f7e2d13"
-TikzPictures = "37f6aa50-8035-52d0-81c2-5a1d08754b2d"
