+++ src/wiring_diagrams/Algebras.jl
@@ -3,8 +3,6 @@
 module WiringDiagramAlgebras
 export oapply, query
 
-using Requires
-
 using ...Present, ...Theories
 using ...Schemas: dom_nums, codom_nums, attr
 using ...CategoricalAlgebra
@@ -269,13 +267,5 @@ make_table(::Type{NamedTuple}, columns, names) =
   NamedTuple{Tuple(names)}(Tuple(columns))
 make_table(::Type{<:DataFrameFallback}, columns, names) =
   make_table(NamedTuple, columns, names)
-
-function __init__()
-  @require DataFrames="a93c6f00-e57d-5684-b7b6-d8193f3e46c0" begin
-    using .DataFrames: DataFrame
-    make_table(::Type{MaybeDataFrame}, columns, names) =
-      make_table(DataFrame, columns, names)
-  end
-end
 
 end
